sbs compare

+++ rely/inference/sbs/sbs_pum.py
@@ -550,7 +550,7 @@ def _value_model_server(args: argparse.Namespace, task_queue: Queue, result_queu
 
     tokenizer = AutoTokenizer.from_pretrained(args.value_model_path, trust_remote_code=True)
     model = AutoModel.from_pretrained(
-        args.value_model_path, num_labels=2, device_map="auto", trust_remote_code=True
+        args.value_model_path, num_labels=2, dtype=torch.bfloat16, device_map="auto", trust_remote_code=True
     )
 
     if tokenizer.pad_token is None:
